re-written with seperate PW validation class
